refactor(tablero): usar mejores patrones
